REFACTOR: local , passConfig 위치 Test로 변경 #6
